Reorganized ResidentialCare boundaries
