+++ LBH.AdultSocialCare.Api/LBH.AdultSocialCare.Api.csproj
@@ -59,10 +59,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="V1\Boundary\ResidentialCarePackageBoundary\Request\" />
-  </ItemGroup>
-
-  <ItemGroup>
     <ProjectReference Include="..\HttpServices\HttpServices.csproj" />
   </ItemGroup>
