+++ LBH.AdultSocialCare.Api/V1/Boundary/ResidentialCare/Request/ResidentialCareAdditionalNeedsCreationRequest.cs
@@ -1,7 +1,7 @@
 using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace LBH.AdultSocialCare.Api.V1.Boundary.Request.ResidentialCare
+namespace LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Request
 {
     public class ResidentialCareAdditionalNeedsCreationRequest
     {
