+++ LBH.AdultSocialCare.Api/V1/Boundary/ResidentialCare/Request/ResidentialCareBrokerageCreationRequest.cs
@@ -1,9 +1,6 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareBrokerageBoundary.Request
+namespace LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Request
 {
     public class ResidentialCareBrokerageCreationRequest
     {
