+++ LBH.AdultSocialCare.Api/V1/Boundary/ResidentialCare/Response/ResidentialCareApprovePackageResponse.cs
@@ -1,10 +1,6 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareAdditionalNeedsBoundary.Response;
 
-namespace LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareApprovePackageBoundary.Response
+namespace LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response
 {
     public class ResidentialCareApprovePackageResponse
     {
