+++ LBH.AdultSocialCare.Api/V1/Boundary/ResidentialCare/Response/ResidentialCareTypeOfStayOptionResponse.cs
@@ -1,9 +1,4 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace LBH.AdultSocialCare.Api.V1.Boundary.Response
+namespace LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response
 {
     public class ResidentialCareTypeOfStayOptionResponse
     {
