+++ LBH.AdultSocialCare.Api/V1/UseCase/Interfaces/ICreateResidentialCarePackageUseCase.cs
@@ -1,6 +1,6 @@
-using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareAdditionalNeedsBoundary.Response;
 using LBH.AdultSocialCare.Api.V1.Domain.ResidentialCarePackageDomains;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
 {
