+++ LBH.AdultSocialCare.Api/V1/UseCase/Interfaces/IGetAllResidentialCareTypeOfStayOptionUseCase.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
 {
