+++ LBH.AdultSocialCare.Api/V1/UseCase/Interfaces/IUpdateResidentialCarePackageUseCase.cs
@@ -4,7 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareAdditionalNeedsBoundary.Response;
+using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.Interfaces
 {
