+++ LBH.AdultSocialCare.Api/V1/UseCase/ResidentialApprovePackageUseCase/Concrete/GetResidentialCareApprovePackageUseCase.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareApprovePackageBoundary.Response;
+using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Gateways.NursingCareApprovePackageGateways;
 using LBH.AdultSocialCare.Api.V1.Gateways.ResidentialCareApprovePackageGateways;
