+++ LBH.AdultSocialCare.Api/V1/UseCase/ResidentialCareUseCases/Interfaces/ICreateResidentialCarePackageUseCase.cs
@@ -1,5 +1,5 @@
 using System.Threading.Tasks;
-using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareAdditionalNeedsBoundary.Response;
+using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response;
 using LBH.AdultSocialCare.Api.V1.Domain;
 using LBH.AdultSocialCare.Api.V1.Domain.ResidentialCarePackageDomains;
