+++ LBH.AdultSocialCare.Api/V1/UseCase/ResidentialCareUseCases/Interfaces/IGetAllResidentialCareTypeOfStayOptionUseCase.cs
@@ -1,6 +1,7 @@
 using LBH.AdultSocialCare.Api.V1.Boundary.Response;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.ResidentialCareUseCases.Interfaces
 {
