+++ LBH.AdultSocialCare.Api/V1/UseCase/ResidentialCareUseCases/Interfaces/IGetResidentialCarePackageUseCase.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Threading.Tasks;
-using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCareAdditionalNeedsBoundary.Response;
+using LBH.AdultSocialCare.Api.V1.Boundary.ResidentialCare.Response;
 
 namespace LBH.AdultSocialCare.Api.V1.UseCase.ResidentialCareUseCases.Interfaces
 {
